generalized query abstraction

+++ src/abc/api/shared/generalizedQuery.ts
@@ -108,31 +108,38 @@ export async function generalizedQuery<T>(
   ctx.cachePerformance.misses =
     ctx.cachePerformance.misses + stalePks.length + newPks.length;
 
+  // PRUNE
+  const removeFromIdx = local.indexedDbPks.filter(i => !serverPks.includes(i));
+  const removeFromVuex = local.vuexPks.filter(i => !serverPks.includes(i));
+  console.log({ removeFromIdx, removeFromVuex });
+
+  if (removeFromVuex.length > 0) {
+    store.commit(
+      `${ctx.vuex.moduleName}/${AbcMutation.ABC_PRUNE_STALE_VUEX_RECORDS}`,
+      { pks: removeFromVuex, vuex: ctx.vuex }
+    );
+  }
+  if (removeFromIdx.length > 0) {
+    ctx.dexieTable.bulkDelete(stalePks).then(() => {
+      // NOTE: this is making the async component (which will be short) not part of the
+      // critical path for this query's completion. This is intended.
+      store.commit(
+        `${ctx.vuex.moduleName}/${AbcMutation.ABC_PRUNE_STALE_IDX_RECORDS}`,
+        { pks: removeFromIdx, vuex: ctx.vuex }
+      );
+    });
+  }
+
   const server: IQueryServerResults<T> = {
     records: serverRecords,
     serverPks,
     newPks,
     cacheHits,
     stalePks,
+    removeFromIdx,
+    removeFromVuex,
     overallCachePerformance: ctx.cachePerformance
   };
-
-  // PRUNE
-  const removeFromIdx = local.indexedDbPks.filter(i => !serverPks.includes(i));
-  const removeFromVuex = local.vuexPks.filter(i => !serverPks.includes(i));
-
-  store.commit(
-    `${ctx.vuex.moduleName}/${AbcMutation.ABC_PRUNE_STALE_VUEX_RECORDS}`,
-    { pks: removeFromVuex, vuex: ctx.vuex }
-  );
-  ctx.dexieTable.bulkDelete(stalePks).then(() => {
-    // NOTE: this is making the async component (which will be short) not part of the
-    // critical path for this query's completion. This is intended.
-    store.commit(
-      `${ctx.vuex.moduleName}/${AbcMutation.ABC_PRUNE_STALE_IDX_RECORDS}`,
-      { pks: removeFromIdx, vuex: ctx.vuex }
-    );
-  });
 
   const response = new AbcResult(ctx, {
     type: "query",
